+++ Twilio/Rest/Api/V2010/Account/IncomingPhoneNumber/LocalInstance.php
@@ -24,6 +24,7 @@ use Twilio\Version;
  * @property \DateTime dateCreated
  * @property \DateTime dateUpdated
  * @property string friendlyName
+ * @property string identitySid
  * @property string phoneNumber
  * @property string origin
  * @property string sid
@@ -65,6 +66,7 @@ class LocalInstance extends InstanceResource {
             'dateCreated' => Deserialize::dateTime(Values::array_get($payload, 'date_created')),
             'dateUpdated' => Deserialize::dateTime(Values::array_get($payload, 'date_updated')),
             'friendlyName' => Values::array_get($payload, 'friendly_name'),
+            'identitySid' => Values::array_get($payload, 'identity_sid'),
             'phoneNumber' => Values::array_get($payload, 'phone_number'),
             'origin' => Values::array_get($payload, 'origin'),
             'sid' => Values::array_get($payload, 'sid'),