+++ Twilio/Rest/Api/V2010/Account/IncomingPhoneNumber/MobileList.php
@@ -155,6 +155,7 @@ class MobileList extends ListResource {
             'VoiceFallbackUrl' => $options['voiceFallbackUrl'],
             'VoiceMethod' => $options['voiceMethod'],
             'VoiceUrl' => $options['voiceUrl'],
+            'IdentitySid' => $options['identitySid'],
         ));
 
         $payload = $this->version->create(