+++ Twilio/Rest/Api/V2010/Account/IncomingPhoneNumberList.php
@@ -172,6 +172,7 @@ class IncomingPhoneNumberList extends ListResource {
             'EmergencyStatus' => $options['emergencyStatus'],
             'EmergencyAddressSid' => $options['emergencyAddressSid'],
             'TrunkSid' => $options['trunkSid'],
+            'IdentitySid' => $options['identitySid'],
         ));
 
         $payload = $this->version->create(