+++ Twilio/Rest/Api/V2010/Account/MessageList.php
@@ -60,6 +60,7 @@ class MessageList extends ListResource {
             'ProviderSid' => $options['providerSid'],
             'ContentRetention' => $options['contentRetention'],
             'AddressRetention' => $options['addressRetention'],
+            'SmartEncoded' => Serialize::booleanToString($options['smartEncoded']),
         ));
 
         $payload = $this->version->create(