+++ Twilio/Tests/Integration/Wireless/V1/SimTest.php
@@ -60,7 +60,8 @@ class SimTest extends HolodeckTestCase {
                 "iccid": "iccid",
                 "e_id": "e_id",
                 "status": "new",
-                "url": "https://wireless.twilio.com/v1/Sims/DEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "url": "https://wireless.twilio.com/v1/Sims/DEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "ip_address": "192.168.1.1"
             }
             '
         ));
@@ -140,7 +141,8 @@ class SimTest extends HolodeckTestCase {
                         "voice_fallback_url": "http://www.example.com",
                         "voice_method": "http_method",
                         "voice_url": "http://www.example.com",
-                        "url": "https://wireless.twilio.com/v1/Sims/DEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                        "url": "https://wireless.twilio.com/v1/Sims/DEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "ip_address": "192.168.1.30"
                     }
                 ],
                 "meta": {
@@ -205,7 +207,8 @@ class SimTest extends HolodeckTestCase {
                 "voice_fallback_url": "http://www.example.com",
                 "voice_method": "http_method",
                 "voice_url": "http://www.example.com",
-                "url": "https://wireless.twilio.com/v1/Sims/DEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "url": "https://wireless.twilio.com/v1/Sims/DEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "ip_address": "192.168.1.30"
             }
             '
         ));